refactor(page): extract useImportExportManager hook (phase 3)
Move file import / drag-drop / scene-handoff + export-thumbnail capture
out of page.tsx into features/import-export/useImportExportManager.ts
(reusing the Phase-1 fileHandling + exportThumbnailOptions modules): the
native/zip file pickers, dropped-file processing + prepare-drag overlay,
Tauri cold-start scene-file launch/handoff, and the thumbnail capture
runner. Late/cross-domain deps (isDesktopRuntime, slicing layer access,
select-all state) flow through a MutableRefObject deps-ref; everything
else is destructure-same-names so TopBar/EmptyScene/modal/save-flow
consumers are unchanged. 28 names returned. Worktree subagent, integrated
+ re-verified here.

page.tsx: 17,869 -> 16,775 lines (-1,094). Verified via `tsc --noEmit`.
Smoke-test file open / drag-drop import / .voxl handoff before merge.
